requirements.txt
Flask==2.2.5
mysql-connector-python==8.0.33
gunicorn==20.1.0
python-dotenv==1.0.1
